refactor to use an internal package
