+++ 500px/mod.ts
@@ -4,7 +4,7 @@
  * @module
  */
 
-import { main } from "./cli.ts";
+import { main } from "./main.ts";
 export * from "./client.ts";
 
 if (import.meta.main) await main();
