--- async/README.md
@@ -1,7 +1,0 @@
-# async ([jsr.io](https://jsr.io/@tugrulates/async))
-
-## Functions
-
-### [pool](https://jsr.io/@tugrulates/async/doc/~/pool)
-
-Maps an array of data to an array of results with concurrency control.
