--- cli/README.md
@@ -1,9 +1,0 @@
-# cli ([jsr.io](https://jsr.io/@tugrulates/cli))
-
-## Classes
-
-### [Config](https://jsr.io/@tugrulates/cli/doc/~/Config)
-
-A string configuration option for CLI.
-
-Options are cached in a key-value store, and not prompted if already set.
