--- cli/deno.json
@@ -1,5 +1,0 @@
-{
-  "name": "@tugrulates/cli",
-  "version": "0.1.0",
-  "exports": "./mod.ts"
-}
