+++ lonely-planet/mod.ts
@@ -1,4 +1,4 @@
-import { main } from "./cli.ts";
+import { main } from "./main.ts";
 export * from "./client.ts";
 
 if (import.meta.main) await main();
