--- module/README.md
@@ -1,12 +1,0 @@
-# module ([jsr.io](https://jsr.io/@tugrulates/readme))
-
-Utilities for managing the modules in this repository.
-
-## Functions
-
-### [generateReadme](https://jsr.io/@tugrulates/readme/doc/~/generateReadme)
-
-Generate a README.md file for a module.
-
-Alternatively, running this file with `deno run` can be used to generate
-README.md files for a list of module.
