--- module/deno.json
@@ -1,5 +1,0 @@
-{
-  "name": "@tugrulates/readme",
-  "version": "0.1.0",
-  "exports": "./mod.ts"
-}
