--- module/mod.ts
@@ -1,7 +1,0 @@
-/**
- * Utilities for managing the modules in this repository.
- *
- * @module
- */
-
-export * from "./readme.ts";
