--- request/README.md
@@ -1,15 +1,0 @@
-# request ([jsr.io](https://jsr.io/@tugrulates/request))
-
-## Classes
-
-### [GraphQLClient](https://jsr.io/@tugrulates/request/doc/~/GraphQLClient)
-
-A GraphQL client for making queries and handling pagination.
-
-### [JsonClient](https://jsr.io/@tugrulates/request/doc/~/JsonClient)
-
-A client for making JSON-based HTTP requests.
-
-### [RequestError](https://jsr.io/@tugrulates/request/doc/~/RequestError)
-
-Represents an error that occurs during a request.
