--- request/mod.ts
@@ -1,2 +1,0 @@
-export * from "./graphql.ts";
-export * from "./request.ts";
